+++ internal/usecase/message/result.go
@@ -4,6 +4,6 @@ import "github.com/WuKongIM/WuKongIM/pkg/wkpacket"
 
 type SendResult struct {
 	MessageID  int64
-	MessageSeq uint32
+	MessageSeq uint64
 	Reason     wkpacket.ReasonCode
 }
